[refactor] Context is not wired through functions
analyzer_context.Context is a Singleton object. It is instantiated
once and never changes. It contains the overall context of the
environment in which CodeChecker analyzer runs. There are many
functions that have too many (7 and above) parameters. One of these
is Context which is unnecessary. Acquiring this object from this
singleton object from the getter function is as efficient as
passing it through parameters.