[analyzer] Introducing AnalysisConfig interface object
The command line flags of "CodeChecker analyze" should be hidden from
the analyzer tools and their driver/manager codes so they are resistant
on changes in command line. Later this object will define the analyzer
plugins' interface. Currently we consider clangsa and clang-tidy as one
plugin. Intuitively we can say that an argparse.Namespace object should
only be used in "cmd" directory.